+++ .vscode/launch.json
@@ -5,6 +5,20 @@
     "version": "0.2.0",
     "configurations": [
       {
+        "name": "Launch Back",
+        "request": "launch",
+        "runtimeArgs": [
+          "run-script",
+          "start"
+        ],
+        
+        "runtimeExecutable": "npm",
+        "skipFiles": [
+          "<node_internals>/**"
+        ],
+        "type": "node"
+      },
+      {
         "name": "Launch via NPM",
         "request": "launch",
         "runtimeArgs": [
