--- db.json
@@ -1,5 +1,0 @@
-{
-  "tasks": [
-    {"title":"Besar a Mauri", "completed": false, "id": 1}
-  ]
-}
