+++ server/data/db.json
@@ -1,0 +1,87 @@
+{
+  "tasks": [
+    {
+      "title": "Z",
+      "completed": false,
+      "id": 5
+    },
+    {
+      "title": "W",
+      "completed": false,
+      "id": 1662594639791
+    },
+    {
+      "title": "F",
+      "completed": false,
+      "id": 1662594716713
+    },
+    {
+      "title": "G",
+      "completed": false,
+      "id": 1662594781279
+    },
+    {
+      "title": "B",
+      "completed": false,
+      "id": 1662594918527
+    },
+    {
+      "title": "5",
+      "completed": false,
+      "id": 1662594972761
+    },
+    {
+      "title": "s",
+      "completed": false,
+      "id": 1662595010757
+    },
+    {
+      "title": "8",
+      "completed": false,
+      "id": 1662595166812
+    },
+    {
+      "title": "H",
+      "completed": false,
+      "id": 1662595200645
+    },
+    {
+      "title": "P",
+      "completed": false,
+      "id": 1662595347610
+    },
+    {
+      "title": "Ñ",
+      "completed": false,
+      "id": 1662595403789
+    },
+    {
+      "title": "J",
+      "completed": false,
+      "id": 1662595435247
+    },
+    {
+      "title": "T",
+      "completed": false,
+      "id": 1662595491960
+    },
+    {
+      "title": "A",
+      "completed": false,
+      "id": 1662595896684
+    },
+    {},
+    {},
+    {},
+    {},
+    {},
+    {},
+    {},
+    {},
+    {},
+    {},
+    {},
+    {}
+  ],
+  "filterType": "all"
+}
