--- server/db.js
@@ -1,9 +1,0 @@
-import { Low, JSONFile } from 'lowdb'
-
-const adapter = new JSONFile('db.json')
-const db = new Low(adapter)
-
-await db.read()
-db.data ||= { tasks: [] }
-
-export default db;
