+++ server/views/home.html
@@ -1,18 +1,11 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-</head>
-<body>
+{% extends "templates/base.html" %}
+{% set title = "Home" %}
 
-  <h1>Todo Tasks</h1>
+{% block content %}
+<h1>Todo Tasks</h1>
 
-  <nav>
-    <a href="/tasks2">Tasks 2</a>
-  </nav>
+<nav>
+  <a href="/tasks2">Tasks 2</a>
+</nav>
 
-</body>
-</html>
+{% endblock %}
