--- server/views/tasks2.ejs
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-</head>
-
-<body>
-  <h1>
-    <a href="/">Todo Tasks</a>
-  </h1>
-
-  <h2>Tasks List</h2>
-
-  <form action="">
-    <input type="text" placeholder="write a task here...">
-    <button>Send</button>
-  </form>
-
-  <ol>
-    <% tasks.forEach(function(task){ %>
-      <li>
-        <%= task.title%>
-      </li>
-    <% }) %>
-  </ol>
-
-</body>
-
-</html>
