+++ server/views/templates/base.html
@@ -4,12 +4,12 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
+  <title>{{ title }}</title>
 </head>
 <body>
-  <div class="todo-app">
 
-  </div>
-  <script  type="module" src="app.js"></script>
+  {% include "templates/navigation.html" %}
+  {% block content %} {% endblock %}
+
 </body>
 </html>
