+++ server/views/templates/navigation.html
@@ -1,0 +1,4 @@
+<nav>
+  <a href="/">Home</a>
+  <a href="/tasks2">Tasks</a>
+</nav>
