--- src/templates/utils.js
@@ -1,6 +1,0 @@
-export default function render(templateString) {
-  const dummyDIV = document.createElement("div")
-
-  dummyDIV.innerHTML = templateString
-  return dummyDIV.firstChild
-}
