+++ templates/utils.js
@@ -1,0 +1,6 @@
+export default function render(templateString) {
+  const dummyDIV = document.createElement("div")
+
+  dummyDIV.innerHTML = templateString
+  return dummyDIV.firstChild
+}
