--- todo.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>TODO App</title>
-</head>
-
-<body>
-  <div class="todo-app"></div>
-
-  <script type="module" src="src/index.js"></script>
-
-</body>
-
-</html>
